requirements.txt
streamlit
pandas
plotly.express
matplotlib.pyplot
numpy